libcollectdclient: Move the header files to a subdirectory.
This makes it easier since the header files depend on one another.
(At least they will in the future.)
