+++ src/libcollectdclient/collectd/client.h
