+++ src/libcollectdclient/collectd/lcc_features.h.in
