+++ src/libcollectdclient/collectd/network_buffer.h
